Implemented serialization to json / zip.
Changed the serialisation method from pickle to json to avoid possible code injection on load.
Made zipping dependent on provided filename, added filename validator. Refactored tests.

Since the vector store uses dataclasses to store its entries, and those are not default json serializable I needed to add a simple json decoder that simply stores the dataclasses dict and a special key that allow to reconstruct the class on deserialization.

Note: If the BaseVectorStore.Entry dataclass gains members that are not json serializable this will break